Conflicts:
	Assets/Scenes/FirstTest.unity

+++ Assets/Scripts/Components/DamageComponent.cs
@@ -5,7 +5,7 @@ public class DamageComponent : MonoBehaviour {
 
 	Transform myOwner;
 
-	int GetWeaponDamage()
+	float GetWeaponDamage()
 	{
 		if ( gameObject.tag == "Player" )
 			return GetComponentInParent<EquipmentComponent>().GetWeaponDamage();
